requirements.txt
airflow==0.6
apache_airflow_providers_imap==3.0.0
boto3==1.25.1
colorama==0.4.5
matplotlib==3.6.0
numpy==1.23.3
pandas==1.5.0
PyYAML==6.0
s3fs==0.4.2
scikit_learn==1.2.0
scorecardpy==0.1.9.2
snowflake==0.0.3
snowflake_connector_python==2.8.0
SQLAlchemy==1.4.41
statsmodels==0.13.2
termcolor==2.1.1
xgboost==1.6.2
